# src/main/resources/api/produit.swagger.yaml
openapi: 3.0.3
info:
  title: API Produit
  description: API REST pour la gestion des produits
  version: 1.0.0
servers:
  - url: http://localhost:8080/api

tags:
  - name: Produits
    description: Gestion des produits

paths:
  /produit/:
    get:
      tags: [ Produit ]
      summary: Lister tous les produits
      operationId: getAllProduits
      responses:
        '200':
          description: Liste des produits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProduitDTO'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [ Produit ]
      summary: Créer un produit
      operationId: createProduit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProduitDTO'
      responses:
        '201':
          description: Produit créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProduitDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /produit/{id}:
    get:
      tags: [ Produit ]
      summary: Récupérer un produit par ID
      operationId: getProduitById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Produit trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProduitDTO'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags: [ Produit ]
      summary: Mettre à jour un produit
      operationId: updateProduit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProduitDTO'
      responses:
        '200':
          description: Produit mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProduitDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [ Produit ]
      summary: Supprimer un produit par ID
      operationId: deleteProduit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Produit supprimé
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /produit/search/by-nom:
    get:
      tags: [ Produit ]
      summary: Rechercher des produits par nom exact
      operationId: findByNomProduit
      parameters:
        - name: nom
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Résultat de recherche
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProduitDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /produit/search/by-nom-contains:
    get:
      tags: [ Produit ]
      summary: Rechercher des produits par nom (contains)
      operationId: findByNomProduitContains
      parameters:
        - name: nom
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Résultat de recherche
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProduitDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /produit/search/by-nom-prix:
    get:
      tags: [ Produit ]
      summary: Rechercher par nom et prix
      operationId: findByNomPrixProduit
      parameters:
        - name: nom
          in: query
          required: true
          schema:
            type: string
        - name: prix
          in: query
          required: true
          schema:
            type: number
            format: double
      responses:
        '200':
          description: Résultat de recherche
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProduitDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /produit/search/by-categorie-id:
    get:
      tags: [ Produit ]
      summary: Rechercher des produits par catégorie (ID)
      operationId: findByCategorieIdCategorie
      parameters:
        - name: idCategorie
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Résultat de recherche
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProduitDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /produit/sort/by-nom:
    get:
      tags: [ Produit ]
      summary: Trier les produits par nom (ASC)
      operationId: findByOrderByNomProduitAsc
      responses:
        '200':
          description: Produits triés
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProduitDTO'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /produit/sort/by-nom-prix:
    get:
      tags: [ Produit ]
      summary: Trier les produits par nom et prix
      operationId: trierProduitsNomsPrix
      responses:
        '200':
          description: Produits triés
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProduitDTO'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  schemas:
    ProduitDTO:
      type: object
      properties:
        idProduit:
          type: integer
          format: int64
        nomProduit:
          type: string
          example: Laptop
        prixProduit:
          type: number
          format: double
          example: 1500.0
        dateCreation:
          type: string
          format: date
        categorie:
          $ref: '#/components/schemas/CategorieDTO'
    CategorieDTO:
      type: object
      properties:
        idCategorie:
          type: integer
          format: int64
          example: 10
        nomCategorie:
          type: string
          example: Informatique
        description:
          type: string

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string

  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    InternalServerError:
      description: Erreur interne du serveur
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

#components:
#  schemas:
#    ProduitDTO:
#      type: object
#      properties:
#        idProduit:
#          type: long
#        nomProduit:
#          type: string
#        prixProduit:
#          type: number
#          format: double
#        categorieId:
#          type: long